Fixed bitmap orientation for Create*Bitmap functions
Fixes upside-down save game screenshots in Siege of Avalon (issue #42)

+++ DDrawCompat/Gdi/VirtualScreen.h
@@ -14,7 +14,7 @@ namespace Gdi
 	{
 		HDC createDc();
 		HBITMAP createDib();
-		HBITMAP createOffScreenDib(DWORD width, DWORD height);
+		HBITMAP createOffScreenDib(LONG width, LONG height);
 		CompatPtr<IDirectDrawSurface7> createSurface(const RECT& rect);
 		void deleteDc(HDC dc);
